Move `stringChecksum` to utils
This is used in a couple of resources now so let's consolidate it.
